feat: Improve usage as a lib
