+++ bin/contentful-migration
@@ -1,5 +1,6 @@
 #!/usr/bin/env node
 
+const usageParams = require('../built/bin/usage-params').default;
 const run = require('../built/bin/cli').default;
 
-run();
+run(usageParams);
